idea1/requirements.txt
torch
numpy
pyyaml
matplotlib

